Add a confirm boc for submit answers in quiz page

+++ Quizzes/quiz/Quiz.js
@@ -107,7 +107,10 @@ function SetRemoveAnswerButtonStatus(questionId, newStatus) {
 }
 //!-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
 //*Handle click on register answers button
-document.getElementById("submit").addEventListener("click", () => {
-    sessionStorage.setItem("quizAnswers", JSON.stringify(answers))
-    window.location.replace(`/Quizzes/quiz/answer-sheet?id=${quizId}`)
+document.getElementById("submit").addEventListener("click", (event) => {
+    const userDecide = confirm("آیا می‌خواهید پاسخ‌های خود را ثبت نمایید؟")
+    if (userDecide) {
+        sessionStorage.setItem("quizAnswers", JSON.stringify(answers))
+        window.location.replace(`/Quizzes/quiz/answer-sheet?id=${quizId}`)
+    }
 })
